Widen font ptrs to ints
Hopefully we'd be overflowing shorts otherwise.

+++ data/font.h
@@ -2,10 +2,10 @@
 #define FONT_H 12
 #ifndef FONTEDITOR
 extern const unsigned char font_data[];
-extern const unsigned short font_ptrs[];
+extern const unsigned int font_ptrs[];
 extern const unsigned int font_ranges[][2];
 #else
 extern unsigned char *font_data;
-extern unsigned short *font_ptrs;
+extern unsigned int *font_ptrs;
 extern unsigned int (*font_ranges)[2];
 #endif
