+++ src/gui/font/FontEditor.h
@@ -25,7 +25,7 @@ private:
 	std::map<String::value_type, std::array<std::array<char, MAX_WIDTH>, FONT_H> > fontPixels;
 
 	std::vector<unsigned char> fontData;
-	std::vector<unsigned short> fontPtrs;
+	std::vector<unsigned int> fontPtrs;
 	std::vector<std::array<unsigned int, 2> > fontRanges;
 
 	ByteString beforeFontData;
@@ -34,18 +34,18 @@ private:
 	ByteString afterFontRanges;
 
 	void ReadDataFile(ByteString dataFile);
-	void WriteDataFile(ByteString dataFile, std::vector<unsigned char> const &fontData, std::vector<unsigned short> const &fontPtrs, std::vector<std::array<unsigned int, 2> > const &fontRanges);
+	void WriteDataFile(ByteString dataFile, std::vector<unsigned char> const &fontData, std::vector<unsigned int> const &fontPtrs, std::vector<std::array<unsigned int, 2> > const &fontRanges);
 	static void PackData(
 			std::map<String::value_type, unsigned char> const &fontWidths,
 			std::map<String::value_type, std::array<std::array<char, MAX_WIDTH>, FONT_H> > const &fontPixels,
 			std::vector<unsigned char> &fontData,
-			std::vector<unsigned short> &fontPtrs,
+			std::vector<unsigned int> &fontPtrs,
 			std::vector<std::array<unsigned int, 2> > &fontRanges);
 	static void UnpackData(
 			std::map<String::value_type, unsigned char> &fontWidths,
 			std::map<String::value_type, std::array<std::array<char, MAX_WIDTH>, FONT_H> > &fontPixels,
 			std::vector<unsigned char> const &fontData,
-			std::vector<unsigned short> const &fontPtrs,
+			std::vector<unsigned int> const &fontPtrs,
 			std::vector<std::array<unsigned int, 2> > const &fontRanges);
 
 	ui::Textbox *currentCharTextbox;
